+++ src/app/components/map-search/map-search.component.html
@@ -1,29 +1,3 @@
-<!-- <div id="search-bars-div">
-  <div class="search-box">
-    <div class="arrow-icon-button">
-      <ion-button class="arrow-button" aria-label="Back" fill="clear" slot="start">
-        <ion-icon name="arrow-back-outline" size="medium" color="medium"></ion-icon></ion-button>
-    </div>
-    <div class="search-bars-container">
-          <ion-searchbar
-            searchIcon="null"
-            class="custom" 
-            placeholder="Choose starting point" 
-            show-clear-button="never" 
-            (ionChange)="onSearchChangeStart($event)">
-          </ion-searchbar>
-    
-        <ion-searchbar
-          searchIcon="null"
-          class="custom" 
-          placeholder="Choose destination point" 
-          show-clear-button="never" 
-          (ionChange)="onSearchChangeDestination($event)">
-        </ion-searchbar>
-    </div>
-  </div>
-</div> -->
-
 <div
   class="rounded-3xl w-full flex text-black bg-white font-bold shadow-lg p-2"
   [ngStyle]="{ 'height': isSearchVisible ? 'auto' : '48px' }"
